Refactor: UserMapper 리펙터링
UserMapper의 위치를 global에서 User 도메인 하위로 이동 및 불필요한 어노테이션 삭제
